requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.0.8
django-allauth==0.51.0
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-post-office==3.6.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
idna==3.4
jsonfield==3.1.0
kombu==5.2.4
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.4
redis==4.3.5
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tinycss2==1.1.1
types-cryptography==3.3.23
tzdata==2022.4
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
